requirements.txt
APScheduler==3.9.1.post1
asgiref==3.7.2
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.1.10
django-apscheduler==0.6.2
django-bootstrap4==23.2
django-ckeditor==6.7.0
django-crispy-forms==1.14.0
django-js-asset==2.1.0
django-reset-migrations==0.4.0
django-staff-required==2020.7.1
feedparser==6.0.10
idna==3.4
Pillow==9.4.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzlocal==5.0.1
urllib3==1.26.16
webencodings==0.5.1
